requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
stripe==9.12.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
whitenoise==6.7.0
